Second review
Fixed
1 - missing functionality for copying a gift, no method in the service, no logic in the controller
2 - When creating and updating a user, there is not enough additional checking and returning a typed error that - User with this email or username is already registered.
What could be improved:

1 - lacks additional checking and return of a typed error that - You have already copied this gift to yourself.
2 - connection address, connection port, it is better to put it in environment variables. env
connect environment variables through a separate config file where we check which variables are passed and which are not, we replace missing variables with default values.
3 - As a recommendation - use transactions for several database accesses within one case.
4 - additional logic: checks, calculations should be performed in the corresponding service. As a rule, in the controller there is only an appeal to the corresponding service, within which all calculations take place.
5 - checking on the user in the controller is duplicated inside the service

Translated with DeepL.com (free version)

+++ src/offers/offers.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -7,7 +8,7 @@ import { CreateOfferDto } from './dto/create-offer.dto';
 import { UpdateOfferDto } from './dto/update-offer.dto';
 import { Offer } from './entities/offer.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Connection, Repository } from 'typeorm';
 import { Wish } from 'src/wishes/entities/wish.entity';
 import { User } from 'src/users/entities/user.entity';
 
@@ -18,6 +19,7 @@ export class OffersService {
     private readonly offerRepository: Repository<Offer>,
     @InjectRepository(Wish)
     private readonly wishRepository: Repository<Wish>,
+    private readonly connection: Connection,
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
   ) {}
@@ -25,34 +27,39 @@ export class OffersService {
   async create(createOfferDto: CreateOfferDto): Promise<Offer> {
     const { userId, wishId, amount, hidden } = createOfferDto;
 
-    const user = await this.userRepository.findOne({ where: { id: userId } });
-    if (!user) {
-      throw new NotFoundException(`User with ID ${userId} not found`);
-    }
+    return await this.connection.transaction(async (manager) => {
+      const user = await manager.findOne(User, { where: { id: userId } });
+      if (!user) {
+        throw new NotFoundException(`User with ID ${userId} not found`);
+      }
 
-    const wish = await this.wishRepository.findOne({ where: { id: wishId } });
-    if (!wish) {
-      throw new NotFoundException(`Wish with ID ${wishId} not found`);
-    }
+      const wish = await manager.findOne(Wish, { where: { id: wishId } });
+      if (!wish) {
+        throw new NotFoundException(`Wish with ID ${wishId} not found`);
+      }
 
-    if (wish.raised >= wish.price) {
-      throw new ConflictException('Funds already raised for this wish');
-    }
+      if (wish.raised >= wish.price) {
+        throw new ConflictException('Funds already raised for this wish');
+      }
 
-    if (amount > wish.price - wish.raised) {
-      throw new ConflictException(
-        'Amount exceeds remaining needed funds for this wish',
-      );
-    }
+      if (amount > wish.price - wish.raised) {
+        throw new ConflictException(
+          'Amount exceeds remaining needed funds for this wish',
+        );
+      }
 
-    const offer = this.offerRepository.create({
-      user,
-      item: wish,
-      amount,
-      hidden,
+      wish.raised += amount;
+      await manager.save(wish);
+
+      const offer = manager.create(Offer, {
+        user,
+        item: wish,
+        amount,
+        hidden,
+      });
+
+      return await manager.save(offer);
     });
-
-    return await this.offerRepository.save(offer);
   }
 
   async findAll(): Promise<Offer[]> {
@@ -71,7 +78,16 @@ export class OffersService {
     return await this.offerRepository.findOne({ where: { id } });
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, userId: number): Promise<void> {
+    const offer = this.offerRepository.findOne({ where: { id } });
+    if (!offer) {
+      throw new NotFoundException(`Offer with ID ${id} not found`);
+    }
+    if ((await offer).user.id != userId) {
+      throw new ForbiddenException(
+        `You do not have permission to delete this wishlist`,
+      );
+    }
     await this.offerRepository.delete(id);
   }
 }
